Data validation done

+++ Controller/StockController.cs
@@ -28,6 +28,8 @@ namespace api_tutorial.Controller
 
         [HttpGet]
         public async Task<IActionResult> GetAll() {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stocks = await _stockRepo.GetAllAsync();
 
             var stockDto = stocks.Select(s => s.ToStockDto());
@@ -37,7 +39,8 @@ namespace api_tutorial.Controller
 
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id) {
-            
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if (stock == null) {
@@ -47,8 +50,10 @@ namespace api_tutorial.Controller
         }
 
         [HttpPost]
-        public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto){
-            
+        public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
+        {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);   
             var stockModel = stockDto.ToStockFromCreateDto();
 
             await _stockRepo.CreateAsync(stockModel);
@@ -58,8 +63,10 @@ namespace api_tutorial.Controller
 
         [HttpPut]
         [Route("{id:int}")]
-        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto){
-            
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
+        {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
             
             if(stockModel == null) {
@@ -70,8 +77,10 @@ namespace api_tutorial.Controller
         }
         [HttpDelete] 
         [Route("{id:int}")]
-        public async Task<IActionResult> Delete([FromRoute] int id) {
-            
+        public async Task<IActionResult> Delete([FromRoute] int id) 
+        {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);    
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if(stockModel == null) {
